Consolidate plugin problem FQN to class resolution to a single place
